Feature/typestate builder
Moving to a public API with a typestate builder. Pipes are now more powerful and can impact the actual nodes rather than just the AST. You can use a basic 2X oversampler through pipes, as well as duplication. More will be built as I dogfood out a couple of apps :)